+++ client/src/components/Navbar/Navbar.js
@@ -55,7 +55,7 @@ class Navbar extends Component {
                         <a className="nav-link" href="/" onClick={() => this.Auth.logout()}>Logout</a>
                     </li>
                     <li className="nav-item">
-                        <Link className="nav-link" to="/ourmushroom">Our Mushrooms</Link>
+                        <Link className="nav-link" to="/ourmushrooms">Our Mushrooms</Link>
                     </li>
                 </ul>
             );
@@ -72,7 +72,7 @@ class Navbar extends Component {
                         <Link className="nav-link" to="/speciescalendar">Calendar</Link>
                     </li>
                     <li className="nav-item">
-                        <Link className="nav-link" to="/ourmushroom">Our Mushrooms</Link>
+                        <Link className="nav-link" to="/ourmushrooms">Our Mushrooms</Link>
                     </li>
                     <li className="nav-item">
                         <Link className="nav-link" to="/adminpage">Admin</Link>
