+++ client/src/index.js
@@ -34,7 +34,7 @@ ReactDOM.render(
             <Route exact path="/login" component={Login} />
             <Route exact path="/signup" component={Signup} />
             <Route exact path="/profile" component={Profile} />
-            <Route exact path="/ourmushroom" component={MushroomPage} />
+            <Route exact path="/ourmushrooms" component={MushroomPage} />
         </div>
     </Router>
     , document.getElementById('root')
